requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain_community
faiss-cpu
google-generativeai
dotenv